+++ app/utils/interfaces/IDownloaderService.ts
