+++ app/utils/interfaces/IPostDownloader.ts
