--- src/App.css
@@ -1,55 +1,0 @@
-.App {
-  font-family: Arial, sans-serif;
-  width: 100%;
-  max-width: 600px;
-  margin: 0 auto;
-  background-color: #f4f4f4;
-  padding: 20px;
-  border-radius: 5px;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
-}
-
-.App h2 {
-  color: #333;
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.App label {
-  display: block;
-  margin-bottom: 10px;
-  color: #666;
-}
-
-.App input[type="text"] {
-  width: 100%;
-  padding: 12px 20px;
-  margin: 8px 0;
-  display: inline-block;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  box-sizing: border-box;
-  font-size: 16px;
-}
-
-.App button {
-  width: 100%;
-  background-color: #4caf50;
-  color: white;
-  padding: 14px 20px;
-  margin: 8px 0;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 16px;
-}
-
-.App button:hover {
-  background-color: #45a049;
-}
-
-.App video {
-  width: 100%;
-  max-height: 500px;
-  margin-top: 20px;
-}
